// src/app/api/models/shoe.ts
/* tslint:disable */
/* eslint-disable */
export interface Shoe {

  /**
   * The brand of the shoe
   */
  brand: string;

  /**
   * The unique ID of the shoe
   */
  id?: number;

  /**
   * The name of the shoe
   */
  name: string;

  /**
   * The price of the shoe
   */
  price: number;

  /**
   * The size of the shoe
   */
  size: string;
}